LPS-49562 Create new local package and move related files to it
